Fix compiler warning
git-svn-id: https://svn.osgeo.org/grass/grass/trunk@10340 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ lib/raster/RGB_ras.c
@@ -25,7 +25,7 @@ int R_RGB_raster(int n, int nrows,
 	_send_char_array(n,grn) ;
 	_send_char_array(n,blu) ;
 	_send_char_array(n,nul ? nul : red) ;
-	z = nul ;
+	z = (int) nul ;
 	_send_int(&z) ;
 
 	return 0;
